trace-cmd: allow python tools to supress warning() and pr_stat()
trace-cmd has a lot of built in warning() and pr_stat() output that is noisy
when you are using the python bindings.  This patch provides overrides for these
two functions and the ability for python scripts to optionally turn them off.
With this patch I can silence all of these extraneous print messages from within
my python script.

Signed-off-by: Josef Bacik <jbacik@fb.com>