trace-cmd: fix report not working with instances
We unconditionally call trace_init_profile in read_data_info, even if we don't
have profile set.  This causes trace-cmd report to fail out if you recorded
using instances because the profile stuff doesn't find the handle for the
instance when it does the record.  So do 2 things

1) Only call trace_init_profile if we actually specified --profile.

2) Make sure to call trace_init_profile for all of the instance handles so that
we get all the info we want and don't bail out.

With this patch I can now do trace-cmd report on a trace.dat file that only had
events in instances.  Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>