trace-cmd: disentangle profile and read from record
trace-record has the bulk of the code we need for creating threads and dealing
with instances.  Rather than going through the trouble of separating all of that
stuff out, just remove the trace-profile and trace-read dependancies from
trace-record.  This way we don't have to make libtracecmd a huge blob with all
of the commands so that linking works.  This is a little ugly, but it can be
made prettier when/if we pull the thread creation and instance handling code out
of trace-record.  Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>