trace-cmd: filter out specific pids
Sometimes it is a good idea to filter out events from the recording processes
and even other pids, for example an external app that links against libtracecmd
that watches syscalls probably doesn't want its own syscalls showing up in the
events.  So add the ability to filter _out_ certain pids for the events that are
enabled.  The recorder threads will add their pids to this list, but they are
only filtered if the library consumer filters out a pid or if the trace-cmd
record user tries to filter anything out.  Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>