trace-cmd: use a list head in trace_hash
We already have a next and prev pointer, but we do it in an awkward way that
doesn't actually let us walk backwards from the start of the list.  Convert the
trace_hash to use a list_head for its buckets so we can do a reverse traversal
of each bucket if we want to.

Signed-off-by: Josef Bacik <jbacik@fb.com>