trace-cmd: add trace_hash_find_reverse
Sometimes we want to FIFO identical entries in the hash table, so add
trace_hash_find_reverse, which will cycle through the hash list in reverse.
This is helpful in our case where we are monitoring block request issues to
match with their corresponding completes, if we issue multiple times to the same
block we want to be able to match them with the correct completes (more or less
anyway).  Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>