trace-cmd: use nonblocking reads for streaming
I noticed while using the streaming infrastructure in trace-cmd that I was
seemingly missing events.  Using other tracing methods I got these events and
record->missed_events was never being set.  This is because the streaming
infrastructure uses blocking reads on the per cpu trace pipe's, which means
we'll wait for an entire pages worth of data to be ready before passing it along
to the recorder.  This makes it impossible to do long term tracing that requires
coupling two different events that could occur on different CPU's, and I imagine
it has been what is screwing up my trace-cmd profile runs on our giant 40 cpu
boxes.  Fix trace-cmd instead to use a nonblocking read with select to wait for
data on the pipe so we don't burn CPU unnecessarily.  With this patch I'm no
longer seeing missed events in my app.  Thanks,

Signed-off-by: Josef Bacik <jbacik@fb.com>